+++ lib/common_utils_import.dart
@@ -3,6 +3,8 @@ export 'dart:collection';
 export 'dart:convert';
 
 export 'package:synchronized/synchronized.dart';
+export 'package:meta/meta.dart';
+export 'package:pedantic/pedantic.dart';
 
 export 'async_utils.dart';
 export 'bool_utils.dart';